update light/dark mode styles
