+++ src/styles/App.sass
@@ -8,19 +8,21 @@
 @mixin light-theme
   --text-color: #3e3e3e
   --bg-color: #fafafa
-  --btn-text-color: #fafafa
-  --btn-bg-color: #303030
-  --btn-content: "Dark"
+  --secondary-color: #dbdbdb
+  --btn-text-color: #f7f7f7
+  --btn-bg-color: #3e3e3e
   --btn-icon: url("/assets/icon-moon.svg") 
+  --logo: url("/assets/logo-easyascii.svg")
 
 // dark theme colors
 @mixin dark-theme
   --text-color: #dadada
-  --bg-color: #252525
+  --bg-color: #303030
+  --secondary-color: #757575
   --btn-text-color: #303030
   --btn-bg-color: #e7e7e7
-  --btn-content: "Light"
   --btn-icon: url("/assets/icon-sun.svg")
+  --logo: url("/assets/logo-easyascii-light.svg")
 
 :root
   @include light-theme
@@ -88,39 +90,10 @@ code
   &:hover,
   &:active,
   &:focus
-    background: var(--text-color)
+    background: var(--secondary-color)
 
 .hide 
   display: none
-
-
-button
-  width: 68px
-  position: absolute
-  top: 1em
-  right: 1em
-  background: var(--btn-bg-color)
-  color: var(--btn-text-color)
-  font-family: $font-family
-  font-size: .775em
-  margin: 0
-  padding: .5em
-  border: none
-  border-radius: 4px
-  cursor: pointer
-  z-index: 1
-
-  &:before 
-    content: var(--btn-icon) 
-    margin-right: 6px
-    line-height: 1
-    display: inline-block
-    vertical-align: middle
-
-  &:after
-    content: var(--btn-content)
-    display: inline-block
-    margin-top: 1px
 
   
 
@@ -128,7 +101,7 @@ button
 header
   z-index: 1
   background: var(--bg-color)
-  padding: 0 1em 1em
+  padding: 1em
   width: 100%
   position: fixed
   top: 0
@@ -140,16 +113,16 @@ header
     display: block
     padding: 0
 
-  img 
+  #logo
     padding-top: 1em
     width: 120px
-    opacity: .85
+    height: 30px
+    background: var(--logo) center no-repeat
 
     @media screen and (max-width: $width-sm)
       padding-left: .9em
 
   nav
-    padding-top: 1em
     position: relative
 
     @media screen and (max-width: $width-sm)
@@ -184,7 +157,24 @@ header
           margin-right: 32px
 
     a:hover
-      box-shadow: 0 2px 0px rgba(0,0,0,.09)
+      box-shadow: 0 2px 0px var(--secondary-color)
+
+    button
+      width: 30px
+      height: 30px
+      background: var(--btn-bg-color)
+      color: var(--btn-text-color)
+      font-family: $font-family
+      font-size: .775em
+      margin: 0
+      padding: .5em
+      border: none
+      border-radius: 100%
+      cursor: pointer
+      line-height: 0
+
+      &:before 
+        content: var(--btn-icon) 
 
   @media screen and (max-width: $width-sm)
     .overlay
@@ -236,8 +226,8 @@ section
         position: absolute
         top: 28px
         left: 0px
-        background: var(--bg-color)
-        color: white
+        background: var(--text-color)
+        color: var(--bg-color)
         font-size: .8em
         font-weight: 400
         padding: 5px 6px 5px 8px
@@ -267,7 +257,7 @@ section
 
     &:before,
     &:after
-      color: var(--text-color)
+      color: var(--secondary-color)
       font-weight: 300
 
     &:before
